pyproject.toml
[tool.poetry]
name = "tyjuliacall"
version = "0.7.4"
description = "Python-Julia interops."
authors = ["Suzhou-Tongyuan <support@tongyuan.cc>"]
packages = [
    { include = "tyjuliacall" },
    { include = "tyjuliasetup" }
]
exclude = [".tongyuan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
julia-numpy = "^0.4.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"


[tool.poetry.group.dev.dependencies]
isort = "5.11.5"
black = "22.12.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
skip = [".tongyuan", ".venv", "logs"]
